rpmsg: Fix compiler warning
GCC issues the following warning:

  ../../../../../../source/rpmsg/pru_virtqueue.c:116:18: warning: comparison of integer expressions of different signedness: ‘int16_t’ {aka ‘short int’} and ‘uint32_t’ {aka ‘long unsigned int’} [-Wsign-compare]
    116 |         if (head > num)

Fix by explicitly using unsigned integers for the comparison.

Negative value should never be passed to pru_virtqueue_add_used_buf
because it is an invalid head. If it is passed, though, making it
unsigned would make it too large, and trigger the early exit with an
error.

+++ source/rpmsg/pru_virtqueue.c
@@ -109,7 +109,7 @@ int16_t pru_virtqueue_add_used_buf(
 	num = vq->vring.num;
 	used = vq->vring.used;
 
-	if (head > num)
+	if ((uint32_t)head > num)
 		return PRU_VIRTQUEUE_INVALID_HEAD;
 
 	/*
